+++ lib/hanami/extensions/view/context.rb
@@ -96,14 +96,12 @@ module Hanami
             # @since 2.1.0
             def initialize( # rubocop:disable Metrics/ParameterLists
               inflector: nil,
-              settings: nil,
               routes: nil,
               assets: nil,
               request: nil,
               **args
             )
               @inflector = inflector
-              @settings = settings
               @routes = routes
               @assets = assets
               @request = request
@@ -267,16 +265,6 @@ module Hanami
             def flash
               request.flash
             end
-
-            private
-
-            # Returns the app's settings.
-            #
-            # @return [Hanami::Settings] the settings
-            #
-            # @api private
-            # @since 2.1.0
-            attr_reader :settings
           end
         end
       end